third_party/blink/web_tests/fast/forms/textarea/textarea-textlength-expected.txt

Test for HTMLTextAreaElement.textLength

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textArea.textLength is 0
PASS textArea.textLength is 4
PASS textArea.textLength is 5
PASS textArea.textLength is 2
PASS textArea.textLength is 2
PASS successfullyParsed is true

TEST COMPLETE
